requirements.txt
altair==4.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
jsonpickle==2.2.0
jsonschema==4.15.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
networkx==2.8.6
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pydeck==0.8.0b1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyvis==0.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.5.0
streamlit==1.12.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
tweepy==4.10.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
zipp==3.8.1
